Validated all endpoints

+++ api/posts/postsRoutes.js
@@ -9,7 +9,7 @@ posts.get('/', (req, res) => {
             res.status(200).json(data);
         })
         .catch((error) => {
-            res.status(500).send({ message: 'boo' })
+            res.status(500).send({ error: "The posts information could not be retrieved." })
         })
 })
 
@@ -17,10 +17,14 @@ posts.get('/:id', (req, res) => {
     const id = req.params.id;
     db.findById(id)
         .then((data) => {
-            res.status(200).json(data);
+            if (data.length ===0){
+                res.status(404).send({ message: "The post with the specified ID does not exist." })
+            } else {
+                res.status(200).json(data);
+            }
         })
         .catch((error) => {
-            res.status(500).send({ message: 'boo' })
+            res.status(500).send({ error: "The post information could not be retrieved." })
         })
 })
 
@@ -28,54 +32,85 @@ posts.get('/:id/comments', (req, res) => {
     const id = req.params.id;
     db.findPostComments(id)
         .then((data) => {
-            res.status(200).json(data);
+            if (data.length === 0){
+                res.status(404).json({ message: "The post with the specified ID does not exist." })
+            } else {
+                res.status(200).json(data);
+            }
+            
         })
         .catch((error) => {
-            res.status(500).send({ message: 'boo' })
+            res.status(500).send({ error: "The comments information could not be retrieved." })
         })
 })
 
 posts.post('/', (req, res) => {
     const post = req.body;
-    db.insert(post)
+    if (post.title && post.contents){
+        db.insert(post)
         .then((data) => {
-            res.status(200).json(data);
+            res.status(201).json(data);
         })
         .catch((error) => {
             res.status(500).send({ message: 'boo' })
         })
+    } else {
+        res.status(400).send({ errorMessage: "Please provide title and contents for the post." })
+    }
+    
 })
 
 posts.post('/:id/comments', (req, res) => {
-    // could retrieve ID from URL and combine it with body to automate post_id key
-    const comment = req.body;
-    db.insertComment(comment)
-        .then((data) => {
-            res.status(200).json(data);
+    const text = req.body.text;
+    const post_id = req.params.id;
+    const comment = {
+        "text": `${text}`,
+        "post_id": `${post_id}`
+    }
+    db.findById(post_id)
+        .then((post) => {
+            if (post.length ===0){
+                res.status(404).send({ message: "The post with the specified ID does not exist." })
+            } else {
+                if (text){
+                    db.insertComment(comment)
+                    .then((data) => {
+                        res.status(201).json(data);
+                    })
+                    .catch((error) => {
+                        res.status(500).send({ error: "There was an error while saving the comment to the database" })
+                    })
+                } else {
+                    res.status(400).send({ errorMessage: "Please provide text for the comment." })
+                }
+            }
         })
         .catch((error) => {
-            res.status(500).send({ message: 'boo' })
+            res.status(500).send({ error: "The post information could not be retrieved." })
         })
 })
 
 
 posts.delete('/:id', (req, res) => {
     const id = req.params.id;
-    db.remove(id)
-        .then((data) => {
-            res.status(200).json(data);
+    db.findById(id)
+    .then((post) => {
+        if (post.length ===0){
+            res.status(404).send({ message: "The post with the specified ID does not exist." })
+        } else {
+            db.remove(id)
+                .then((data) => {
+                    res.status(200).json(data);
+                })
+                .catch((error) => {
+                    res.status(500).send({ error: "The post could not be removed" })
         })
-        .catch((error) => {
-            console.log("ERRRRR", error)
-            res.status(500).send({ message: 'boo' })
-        })
+        }
+    })
+    .catch((error) => {
+        res.status(500).send({ error: "The post information could not be retrieved." })
+    })
 })
-
-
-
-
-
-
 
 
 module.exports = posts;
